+++ src/Web/Dispatcher.Web/Controllers/JobsController.cs
@@ -2,7 +2,6 @@
 {
     using System.Threading.Tasks;
 
-    using Dispatcher.Common;
     using Dispatcher.Data.Models;
     using Dispatcher.Services.Contracts;
     using Dispatcher.Services.Data.Contracts;
@@ -11,21 +10,25 @@
     using Microsoft.AspNetCore.Identity;
     using Microsoft.AspNetCore.Mvc;
 
+    using static Dispatcher.Common.GlobalConstants.Job;
     using static Dispatcher.Common.GlobalConstants.PageEntities;
 
     public class JobsController : Controller
     {
         private readonly IJobService jobService;
         private readonly IStringValidatorService stringValidatorService;
+        private readonly IPermissionsValidatorService permissionsValidator;
         private readonly UserManager<ApplicationUser> userManager;
 
         public JobsController(
             IJobService jobService,
             IStringValidatorService stringValidatorService,
+            IPermissionsValidatorService permissionsValidator,
             UserManager<ApplicationUser> userManager)
         {
             this.jobService = jobService;
             this.stringValidatorService = stringValidatorService;
+            this.permissionsValidator = permissionsValidator;
             this.userManager = userManager;
         }
 
@@ -40,13 +43,13 @@
         public async Task<IActionResult> Create(JobInputModel input)
         {
             if (!this.ModelState.IsValid
-                || !this.stringValidatorService.IsStringValidDecoded(input.JobBody, GlobalConstants.DefaultBodyStringMinLength))
+                || !this.stringValidatorService.IsStringValidDecoded(input.JobBody, BodyMinLength))
             {
                 return this.View();
             }
 
-            var user = await this.userManager.GetUserAsync(this.User);
-            await this.jobService.CreateAsync(input, user.Id);
+            var userId = this.userManager.GetUserId(this.User);
+            await this.jobService.CreateAsync(input, userId);
 
             return this.RedirectToAction(nameof(this.AllJobs));
         }
@@ -54,7 +57,13 @@
         [Authorize]
         public IActionResult Edit(int id)
         {
+            if (!this.HasPermission(id))
+            {
+                return this.RedirectToAction("Error", "Home");
+            }
+
             var job = this.jobService.GetById<EditJobInputModel>(id);
+
             return this.View(job);
         }
 
@@ -62,13 +71,19 @@
         [Authorize]
         public async Task<IActionResult> Edit(EditJobInputModel input, int id)
         {
+            if (!this.HasPermission(id))
+            {
+                return this.RedirectToAction("Error", "Home");
+            }
+
             if (!this.ModelState.IsValid
-                || !this.stringValidatorService.IsStringValidDecoded(input.JobBody, GlobalConstants.DefaultBodyStringMinLength))
+                || !this.stringValidatorService.IsStringValidDecoded(input.JobBody, BodyMinLength))
             {
                 return this.View(input);
             }
 
             await this.jobService.UpdateAsync<EditJobInputModel>(input, id);
+
             return this.RedirectToAction(nameof(this.Job), new { id });
         }
 
@@ -87,12 +102,20 @@
         public IActionResult Job(int id)
         {
             var job = this.jobService.GetById<SigleJobViewModel>(id);
+
             return this.View(job);
         }
 
+        [Authorize]
         public async Task<IActionResult> Delete(int id)
         {
+            if (!this.HasPermission(id))
+            {
+                return this.RedirectToAction("Error", "Home");
+            }
+
             await this.jobService.DeleteAsync(id);
+
             return this.RedirectToAction(nameof(this.AllJobs));
         }
 
@@ -114,6 +137,15 @@
             };
 
             return this.View(jobs);
+        }
+
+        private bool HasPermission(int dataId)
+        {
+            var hasPermission = this.permissionsValidator.HasPermission(
+              this.jobService.GetCreatorId(dataId),
+              this.userManager.GetUserId(this.User));
+
+            return hasPermission.Result;
         }
     }
 }
